web: community queries, one file per domain
server/queries.ts held every read the community makes: 1,364 lines, patterns beside decks beside territories beside notifications. It is a barrel now over server/queries/{patterns,decks,workshop,moderation,users,notifications}.ts, with the author join and the counted sub-selects they all lean on in shared.ts. A script cut the file at its top-level declarations (comments riding with the declaration below them), assigned exports by name, sent each private helper to the one domain that reaches it or to shared when several do, and computed each file imports from what its body names; the barrel re-exports exactly the names the file exported before, so the fifty-odd importers did not change. Bodies are verbatim; the twenty smoke suites and the build pass.
